Include a default application in the docker image
This places a default pvw application in `/pvw` in sandtank-dev, and
it gets copied down to the runtime. The client is installed and placed
inside `/pvw/www`.

This will allow a simpler command for starting a docker image on a
local computer, via `docker run -p 9000:80 -it hydroframe/sandtank:web-service`
when the docker image gets pushed. It will use the default application.

The default `/pvw` application can also be overridden by mounting another
one on top of it in the docker run command.

The build scripts were also updated so that they hopefully will work
when ran from other directories.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>