Add GitHub Actions to build docker images
This builds the docker images one at a time, and then pushes the web
service image at the end when it finishes.

For pull requests, it still builds all of the images, but it will not
push the web service image. It will only push the web service image when
the pull request has been merged.

The docker tag is "latest" if a PR was merged. If a git tag was pushed,
then the docker tag is the git tag.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>