ci(docker): push only on release
Rather than pushing a docker image when a git tag is pushed (which also doesn't seem to work when
semantic-release pushes a new tag), push a docker image when a new release is made. This will update
"latest", and it will also push a new tag with the version number.